requirements.txt
GitPython==2.1.0
mock==2.0.0
nose==1.3.7
python-hglib==2.2
xmltodict==0.10.2
coverage==4.2
